lit to module
